requirements.txt
aioredis==1.1.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
cffi==1.11.5
channels==2.1.2
channels-redis==2.2.1
constantly==15.1.0
cryptography==2.3
daphne==2.2.1
Django==2.1
django-bootstrap==0.2.4
django-bootstrap4==0.0.6
django-smartfields==1.0.9
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytz==2018.5
pywin32==223
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
txaio==18.7.1
zope.interface==4.5.0
